+++ src/s_sound.h
@@ -121,7 +121,6 @@ void S_ResumeMusic(void);
 //
 void S_UpdateSounds(void);
 
-void S_SetMusicVolume(int volume);
 void S_LowerMusicVolume(void);
 void S_RestoreMusicVolume(void);
 void S_SetSfxVolume(int volume);